make generic for partials too
